Move appeals issue card to shared folder
