--- src/applications/appeals/995/tests/components/IssueCard.unit.spec.jsx
@@ -1,146 +1,0 @@
-import React from 'react';
-import { expect } from 'chai';
-import { render, fireEvent, waitFor } from '@testing-library/react';
-import sinon from 'sinon';
-
-import { $, $$ } from 'platform/forms-system/src/js/utilities/ui';
-
-import { IssueCardContent, IssueCard } from '../../components/IssueCard';
-import { errorMessages } from '../../constants';
-
-import { SELECTED } from '../../../shared/constants';
-
-const getContestableIssue = (id, selected) => ({
-  ratingIssueSubjectText: `issue-${id}`,
-  description: 'blah',
-  ratingIssuePercentNumber: id,
-  approxDecisionDate: `2021-01-${id}`,
-  [SELECTED]: selected,
-});
-const getAdditionalIssue = (id, selected) => ({
-  issue: `new-issue-${id}`,
-  decisionDate: `2021-02-${id}`,
-  [SELECTED]: selected,
-});
-
-describe('<IssueCard>', () => {
-  const getProps = ({
-    showCheckbox = true,
-    onChange = () => {},
-    onEdit = null,
-  } = {}) => ({
-    id: 'id',
-    index: 0,
-    options: { appendId: 'z' },
-    showCheckbox,
-    onChange,
-    onEdit,
-  });
-
-  it('should render a Contestable issue', () => {
-    const props = getProps();
-    const issue = getContestableIssue('10');
-    const { container } = render(<IssueCard {...props} item={issue} />);
-    expect($('h4', container)).to.exist;
-    expect($('input[type="checkbox"]', container)).to.exist;
-    expect($('.widget-title', container).textContent).to.eq('issue-10');
-    expect($('.widget-title.dd-privacy-hidden', container)).to.exist;
-    expect($('.widget-content', container).textContent).to.contain('blah');
-    expect($('.widget-content', container).textContent).to.contain(
-      'Current rating: 10%',
-    );
-    expect($('.widget-content', container).textContent).to.contain(
-      'Decision date: January 10, 2021',
-    );
-    expect($$('a.edit-issue-link').length).to.equal(0);
-  });
-  it('should render an Additional issue', () => {
-    const props = getProps({ onEdit: () => {} });
-    const issue = getAdditionalIssue('22');
-    const { container } = render(<IssueCard {...props} item={issue} />);
-    expect($('input[type="checkbox"]', container)).to.exist;
-    expect($('.widget-title', container).textContent).to.eq('new-issue-22');
-    expect($('.widget-content', container).textContent).to.contain(
-      'Decision date: February 22, 2021',
-    );
-    expect($('a.edit-issue-link', container)).to.exist;
-  });
-  it('should render a selected issue with appendId included', () => {
-    const props = getProps();
-    const issue = getContestableIssue('01', true);
-    const { container } = render(<IssueCard {...props} item={issue} />);
-    const checkbox = $('input[type="checkbox"]', container);
-    expect(checkbox).to.exist;
-    expect(checkbox.id).to.equal('id_0_z'); // checks appendId
-    expect(checkbox.checked).to.be.true;
-  });
-
-  it('should render a selected additional issue without the checkbox or edit button', () => {
-    const props = getProps({ showCheckbox: false });
-    const issue = getAdditionalIssue('03', true);
-    const { container } = render(<IssueCard {...props} item={issue} />);
-    expect($$('input[type="checkbox"]', container).length).to.equal(0);
-    expect($$('.edit-issue-link', container).length).to.equal(0);
-  });
-
-  it('should call onChange when the checkbox is toggled', async () => {
-    const onChange = sinon.spy();
-    const props = getProps({ onChange });
-    const issue = getContestableIssue('01', true);
-    const { container } = render(<IssueCard {...props} item={issue} />);
-    const element = $('input[type="checkbox"]', container);
-
-    // "Click" the option
-    fireEvent.click(element);
-    // Check that it changed
-    await waitFor(() => {
-      expect(onChange.called).to.be.true;
-      expect(onChange.firstCall.args[1].target.checked).to.be.true;
-    });
-  });
-});
-
-describe('<IssueCardContent>', () => {
-  it('should render an empty div', () => {
-    const { container } = render(<IssueCardContent />);
-    expect($('.widget-content-wrap', container).textContent).to.equal('');
-  });
-  it('should render Contestable issue content', () => {
-    const issue = getContestableIssue('20');
-    const { container } = render(<IssueCardContent {...issue} />);
-
-    const text = $('.widget-content-wrap', container).textContent;
-    expect(text).to.contain('blah');
-    expect(text).to.contain('Current rating: 20%');
-    expect(text).to.contain('Decision date: January 20, 2021');
-    expect($$('a.edit-issue-link', container).length).to.equal(0);
-  });
-  it('should render AdditionalIssue content', () => {
-    const issue = getAdditionalIssue('21');
-    const { container } = render(<IssueCardContent {...issue} />);
-    expect($('.widget-content-wrap', container).textContent).to.contain(
-      'Decision date: February 21, 2021',
-    );
-  });
-  it('should render API loaded issue with invalid date message', () => {
-    const issue = {
-      ...getContestableIssue('20'),
-      approxDecisionDate: '2020-?-?',
-    };
-    const { container } = render(<IssueCardContent {...issue} />);
-
-    expect($('.widget-content-wrap', container).textContent).to.contain(
-      `Decision date: ${errorMessages.cardInvalidDate}`,
-    );
-  });
-  it('should render AdditionalIssue invalid date message', () => {
-    const issue = {
-      ...getAdditionalIssue('21'),
-      decisionDate: '2021-?-?',
-    };
-    const { container } = render(<IssueCardContent {...issue} />);
-    expect($('.widget-content-wrap', container).textContent).to.contain(
-      `Decision date: ${errorMessages.cardInvalidDate}`,
-    );
-  });
-});
