+++ src/applications/appeals/996/components/ContestableIssuesWidget.jsx
@@ -7,9 +7,9 @@ import { VaModal } from '@department-of-veterans-affairs/component-library/dist/
 import set from 'platform/utilities/data/set';
 import { setData } from 'platform/forms-system/src/js/actions';
 
-import { IssueCard } from './IssueCard';
 import { APP_NAME } from '../constants';
 
+import { IssueCard } from '../../shared/components/IssueCard';
 import {
   ContestableIssuesLegend,
   NoIssuesLoadedAlert,
